tracking precision difference between Debug and Release builds in quad-double

+++ include/universal/number/qd/mathext/horners.hpp
@@ -25,7 +25,7 @@ namespace sw { namespace universal {
 
         for (int i = n - 1; i >= 0; --i) {
             r *= x;
-            r += coefficients[i];
+            r += coefficients[static_cast<unsigned>(i)];
         }
 
         return r;
@@ -43,14 +43,14 @@ namespace sw { namespace universal {
     inline qd polyroot(const std::vector<qd>& c, const qd& x0, int max_iter, double threshold = 1e-32) {
         if (threshold == 0.0) threshold = qd_eps;
 
-        int n = c.size() - 1;
+        int n = static_cast<int>(c.size()) - 1;
         double max_c = std::abs(double(c[0]));
         // Compute the coefficients of the derivatives
         std::vector<qd> derivatives{ c };
         for (int i = 1; i <= n; ++i) {
             double v = std::abs(double(c[static_cast<unsigned>(i)]));
             if (v > max_c) max_c = v;
-            derivatives[i - 1] = c[static_cast<unsigned>(i)] * static_cast<double>(i);
+            derivatives[static_cast<unsigned>(i - 1)] = c[static_cast<unsigned>(i)] * static_cast<double>(i);
         }
         threshold *= max_c;
